requirements.txt
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
mysqlclient==2.2.4
python-decouple==3.8
gunicorn==21.2.0
